Fix alignment of power-use values.
Previously, padding was used to try to get these the correct height.
However, the height of the div ended up being different between the
'Power Now' and 'Used Today' values because of different font sizes used
within the Used Today div (for the 'kWh' unit).  The incorrect height of
these lines was coming from the emoncms 'bootstrap' CSS setting
line-height to 20px.

Newer bootstrap.css replaces this with a multiplier of approx 1.4
instead of a fixed values, but updating bootstrap CSS is a significant
change as it is substantially different than when first imported. and
will alter the layout of all pages.  Instead, just set the line height
we'd like for the 'power-use' class to correct this problem, and then
remove the padding since it is no longer needed.

Fixes #22.